README.me
Make a class called String that has a sentence as attributes (of type of
String) and its length. In main the object will be created and the user will be prompted to enter
a phrase that can be one word or several words separated by a space in
blank and through the set methods, the phrase and length will be saved as
automatically based on the length of the entered phrase. You must also implement the
following methods:
• ShowVowels() method, you must count the number of vowels that the
entered phrase.
• InvestPhrase() method, it must invert the entered sentence and display it on the screen.
For example: Input: "white house", Output: "acnalb asac".
Method timesRepeated(String letter), will receive a character entered by the user
and count how many times the character is repeated in the phrase, for example:
•
Input: phrase = "white house". Output: The character 'a' is repeated 4 times.
•
•
Method compareLength(String phrase), you must compare the length of the phrase
that composes the class with another new sentence entered by the user.
Method joinphrases(String phrase), must join the phrase contained in the String class
with a new phrase entered by the user and display the resulting phrase.
• Method replace(String letter), it must replace all the letters “a” that are
found in the sentence, by some other character selected by the user and
display the resulting sentence.
• Method contains(String letter), it must check if the phrase contains a letter that
enter the user and return true if it contains it and false if not