README.me
Check digit. Create a NIF class that will be used to maintain DNIs with your
corresponding letter (NIF). The attributes will be the ID number (long integer) and the
corresponding letter (String or char). You will have the following methods:
• Methods getters and setters for the ID number and the letter.
• CreateNif() method: asks the user for the DNI and with that DNI calculates the letter that
correspond to. Once calculated, it assigns the letter that corresponds to it according to the
calculation result.
• Method show(): that allows us to show the NIF (eight digits, a hyphen and the letter
in capital letters; for example: 00395469-F).
The letter corresponding to the check digit will be calculated through a method that
works as follows: To calculate the letter, take the remainder of dividing the
ID number by 23 (the result must be a number between 0 and 22). The method must
search in an array (vector) of characters for the position that corresponds to the rest of the
division to get the corresponding letter. The character table is as follows:
POSITION
LYRICS
0T
1R
2W
3A
4G
5M
6 and
7F
8P
9D
10x
11B
12N
13J
14Z
15S
16 Q
17V
6pm
19L
20C
21K
22E